requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.16.4
    # via flask-migrate
bcrypt==4.3.0
    # via flask-bcrypt
blinker==1.9.0
    # via flask
brotli==1.1.0
    # via fonttools
cffi==1.17.1
    # via weasyprint
click==8.2.1
    # via flask
cssselect2==0.8.0
    # via weasyprint
dnspython==2.7.0
    # via pymongo
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.1
    # via
    #   teacher-schedule-dashboard (pyproject.toml)
    #   flask-bcrypt
    #   flask-login
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-testing
    #   flask-wtf
flask-bcrypt==1.0.1
    # via teacher-schedule-dashboard (pyproject.toml)
flask-login==0.6.3
    # via teacher-schedule-dashboard (pyproject.toml)
flask-migrate==4.1.0
    # via teacher-schedule-dashboard (pyproject.toml)
flask-sqlalchemy==3.1.1
    # via
    #   teacher-schedule-dashboard (pyproject.toml)
    #   flask-migrate
flask-testing==0.8.1
    # via teacher-schedule-dashboard (pyproject.toml)
flask-wtf==1.2.2
    # via teacher-schedule-dashboard (pyproject.toml)
fonttools==4.59.0
    # via weasyprint
gunicorn==23.0.0
    # via teacher-schedule-dashboard (pyproject.toml)
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
natsort==8.4.0
    # via teacher-schedule-dashboard (pyproject.toml)
numpy==2.3.2
    # via
    #   teacher-schedule-dashboard (pyproject.toml)
    #   pandas
openpyxl==3.1.5
    # via teacher-schedule-dashboard (pyproject.toml)
packaging==25.0
    # via gunicorn
pandas==2.3.1
    # via teacher-schedule-dashboard (pyproject.toml)
pillow==11.3.0
    # via weasyprint
psycopg2-binary==2.9.10
    # via teacher-schedule-dashboard (pyproject.toml)
pycparser==2.22
    # via cffi
pydyf==0.11.0
    # via weasyprint
pymongo==4.13.2
    # via teacher-schedule-dashboard (pyproject.toml)
pyphen==0.17.2
    # via weasyprint
python-dateutil==2.9.0.post0
    # via pandas
python-decouple==3.8
    # via teacher-schedule-dashboard (pyproject.toml)
python-dotenv==1.1.1
    # via teacher-schedule-dashboard (pyproject.toml)
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.41
    # via
    #   alembic
    #   flask-sqlalchemy
tinycss2==1.4.0
    # via
    #   cssselect2
    #   weasyprint
tinyhtml5==2.0.0
    # via weasyprint
typing-extensions==4.14.1
    # via
    #   alembic
    #   sqlalchemy
tzdata==2025.2
    # via pandas
weasyprint==66.0
    # via teacher-schedule-dashboard (pyproject.toml)
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
    #   tinyhtml5
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
wtforms==3.2.1
    # via flask-wtf
zopfli==0.2.3.post1
    # via fonttools
